+++ src/com/walmartlabs/lacinia/schema.clj
@@ -17,6 +17,7 @@
              cond-let ->TaggedValue is-tagged-value? extract-value extract-type-tag]]
     [com.walmartlabs.lacinia.resolve :refer [ResolverResult resolve-as combine-results]]
     [clojure.string :as str]
+    [clojure.set :refer [difference]]
     [clojure.spec.test.alpha :as stest])
   (:import
     (com.walmartlabs.lacinia.resolve ResolverResultImpl)
@@ -243,20 +244,6 @@
   ([message data]
    (merge {:message message} data)))
 
-(defn ^:private schema-reference?
-  "A reference to another type in the schema. These can always be either
-  a keyword or a symbol."
-  [v]
-  (or (keyword? v)
-      (symbol? v)))
-
-(defn ^:private named?
-  "True if a string, symbol or keyword."
-  [v]
-  (or (string? v)
-      (symbol? v)
-      (keyword? v)))
-
 (defn ^:private graphql-identifier?
   "Expects a conformed value (turns out to be a Map$Entry such as `[:keyword :foo]`)
   and validates that the value is a valid GraphQL identifier."
@@ -362,15 +349,9 @@
                    ::mutations
                    ::subscriptions]))
 
-(s/def ::resolver
-  (s/fspec :args (s/cat ::context (s/keys)
-                        ::arguments (s/map-of keyword? any?)
-                        ::value any?)
-           :ret any?))
-
-(s/def ::default-field-resolver
-  (s/fspec :args (s/cat :field keyword?)
-           :ret ::resolver))
+;; Again, this can be fleshed out once we have a handle on defining specs for
+;; functions:
+(s/def ::default-field-resolver fn?)
 
 (s/def ::compile-options (s/keys :opt-un [::default-field-resolver]))
 
@@ -973,24 +954,54 @@
   [schema object options]
   (verify-fields-and-args schema object)
   (doseq [interface-name (:implements object)
-          :let [interface (get schema interface-name)]
+          :let [interface (get schema interface-name)
+                type-name (:type-name object)]
           [field-name interface-field] (:fields interface)
-          :let [object-field (get-in object [:fields field-name])]]
-
-    ;; TODO: I believe we are missing a check that arguments on the field are
-    ;; compatible with arguments on the interface field.
+          :let [object-field (get-in object [:fields field-name])
+                interface-field-args (:args interface-field)
+                object-field-args (:args object-field)]]
 
     (when-not object-field
       (throw (ex-info "Missing interface field in object definition."
-                      {:object (:type-name object)
+                      {:object type-name
                        :field-name field-name
                        :interface-name interface-name})))
 
     (when-not (is-assignable? schema interface-field object-field)
       (throw (ex-info "Object field is not compatible with extended interface type."
-                      {:object (:type-name object)
+                      {:object type-name
                        :interface-name interface-name
-                       :field-name field-name}))))
+                       :field-name field-name})))
+
+    (when interface-field-args
+      (doseq [interface-field-arg interface-field-args
+              :let [[arg-name interface-arg-def] interface-field-arg
+                    object-field-arg-def (get object-field-args arg-name)]]
+
+        (when-not object-field-arg-def
+          (throw (ex-info "Missing interface field argument in object definition."
+                          {:object type-name
+                           :field-name field-name
+                           :interface-name interface-name
+                           :argument-name arg-name})))
+
+        (when-not (is-assignable? schema interface-arg-def object-field-arg-def)
+          (throw (ex-info "Object field's argument is not compatible with extended interface's argument type."
+                          {:object type-name
+                           :interface-name interface-name
+                           :field-name field-name
+                           :argument-name arg-name})))))
+
+    (when-let [additional-args (seq (difference (into #{} (keys object-field-args))
+                                                (into #{} (keys interface-field-args))))]
+      (doseq [additional-arg-name additional-args
+              :let [arg-kind (get-in object-field-args [additional-arg-name :type :kind])]]
+        (when (= arg-kind :non-null)
+          (throw (ex-info "Additional arguments on an object field that are not defined in extended interface cannot be required."
+                          {:object type-name
+                           :interface-name interface-name
+                           :field-name field-name
+                           :argument-name additional-arg-name}))))))
 
   (update object :fields #(reduce-kv (fn [m field-name field]
                                        (assoc m field-name